Created test suite and renamed test files. Tests are now system agnostic and should run without having to modify paths
